Pavel/side menu (#1364)
* wip

* side-menu

* cleanup

* more cleanup

* side menu keyboard cotrol

* same

* more cleanup
